ib_srp: fix unload srp connected devices during stop flow in openibd
In openibd stop flow fixed resolving of connected, using srp,
scsi devices, changed parameters to be sent to iscsi_id utility
and added checking/verification that scsi_id utility exist on server.
In case and it is (scsi_id) not exist, appropriate message will be printed.
Stop flow in this case will failed, because mlx4_ib driver
will be reported as in use (because of not disconnected srp sessions)
and openibd will exit with error.

Issue: 172357

Signed-off-by: Alex Tabachnik <alext@mellanox.com>

Change-Id: I5f8e9c741b87d064ef8ff7cb14071f88225693bc