mlx4_en: ignore init and shutdown hypervisor events
Messages in dmesg are harmless.
They are MLX4_DEV_EVENT_SLAVE_INIT and MLX4_DEV_EVENT_SLAVE_SHUTDOWN
events used by Hypervisor to inform the PPF IB driver that
IB paravirtualization must be initialized/destroyed for a slave.
The fix is to ignore these events in the mlx4_en driver
(they are in fact ignored, aside from the printout).

issue: 174011

Change-Id: I81009f59611ade3c72513fbc6b82780ae80fad74
Signed-off-by: Eugenia Emantayev <eugenia@mellanox.co.il>