mlx4_core: Add support for different value formats in BDF-Val params
The patch sets the convention of using probe_vf and probe_vf_str params.
Each of the params may be used it in one of two ways:
1. The original way: num_vfs=Val
    - Val: The num of VFs to set for all devices (any number in the range 0-64)
2. The new way: num_vfs=bb:dd.f-Val,...
    - 'bb:dd.f' : Strict hexadecimal format defines the BDF (e.g. '04:00.0').
	- Val : The num of VFs to set for the defined BDF (range 0-64).
The same format is used for both probe_vf and probe_vf_str.

The patch also sets the convention of using port_type_array param.
The param may be used one of two ways:
1. The original way: num_vfs=t1,t2
    - t1 & t1: The port types of port 1 and port 2 respectively for all devices
	  (t1 and t2 may be any number in the range 0-3)
2. The new way: num_vfs=bb:dd.f-t1;t2,...
    - 'bb:dd.f' : Strict hexadecimal format defines the BDF (e.g. '04:00.0').
    - t1 & t1: The port types for the defined BDF (range 0-3)

Notice that the seperator between t1 and t2 is not the same in the 2 ways.

issue: 175702
Change-Id: Ic78ad13cb68a59b8137f884e6e0a499847237456
Signed-off-by: Moshe Lazer <moshel@mellanox.com>