mlx4_en: help disperse interrupts workload in TX flow
Make sure that at least once in 2048 completions, the CQ will be armed so an
EQE will will be issued and eventually an interrupt request. This is necessary
since the irq balancing code takes effect after the handler is run and we want
the irq balance daemon configurations to take effect.

issue: 130864

Change-Id: Ie631b9078ed56c62355accb82aa500f258281ff4
Signed-off-by: Eugenia Emantayev <eugenia@mellanox.co.il>