ib/ipoib: protect client data with spin-lock.
protect the client data whenever 2 process are trying to get it into
ipoib_remove_port.
it is possible that the kernel calls remove_one at the same time ipoib cals it
(whenever it failed to ipoib_add_port)
the solution is to protect the ib_get_client_data call with spin_lock.

issue:173881
issue:177888

Change-Id: I17fbb3d03ee541b76de79d062ac0b4088b43f008
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>