mlx4_ib: assign all dev numbers in low-level driver when dev_assign_str param is active
When the dev_assign_str module parameter is in use, any BDF ConnectX
devices which are not included in the parameter must be assigned device
numbers by the low-level mlx4_ib driver.  Otherwise, the ib core layer
may assign numbers which were already allocated in the dev_assign_str
parameter, leading to conflicts and Oopses.

issue: 172086

Change-Id: I3e81da7201598d903432d475031f5fdd8c456e82
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>