mlx4_core: Ignore bf flag in mlx4_qp_reserve_range when count is too large
IPoIB requests allocating a QP range = roundup_power_of_2(num_cores + 1).
When the number of active cores is 64, the qp count is 128.

Currently, this fails a max-count check for allocating BF-enabled QPs, due
to requirements for an ETH port (when the range is allocated we do not yet
know, in VPI, which port the QPs are intended for, so we need to satisfy
the strictest requirements).

Currently, the only client for this capability which requests large counts
is IPoIB TSS/RSS -- which only runs on IB ports.
Since for IB ports BF-enabled QPs need no special handling, we can safely
turn off the BF-enablement flag when allocating the QP range when the
count is too large.

If/when there are other clients for the QPG capability (allocating a
consecutive range of bf-enabled QPs) for IB apps which can run over RoCE,
this patch will need to be expanded to turn off BF for the QP in the
RESET-to-INIT qp transition if we are running in RoCE and the QP number
is not appropriate for BF over an ETH port.

issue: 187044

Change-Id: Ib493fe428484603abc28afa72403a768bf673acd
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>