mlx4_core: sysfs, fix usage of log_num_mtt module parameter
When was auto calculated based on RAM size it wrongly includes
also log_mtts_per_seg.

It's wrong in 2 ways:
First, log_mtts_per_seg should be added by the application itself,
no reason to a have total in log_num_mtt itself.
Second, in case that an extra  NIC exists it may get an invalid value as it
depends on a larger value.
Specifically, it may cause an overflow and later leads on to kernel panic via
mlx4_buddy_init.

issue: 193043
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>

Change-Id: I3f5207cb2335c4f974c5a2fb6f092644e088181a