net/dcb: Add an optional QCN attribute [net-next:qcn]
As specified in 802.1Qau spec.
This patch adds this optional attribute to DCB netlink.
To use it, drivers should implement and register the callbacks
ieee_getqcn, ieee_setqcn and ieee_getqcnstats.

QCN attribute holds a set of parameters for management,
and a set of statistics to provide informative data.

issue: 370621

Change-Id: I1d4d64949a60e345dda760d41945103e5ceb513e
Signed-off-by: Shani Michaeli <shanim@mellanox.com>