IB/mlx5: Improve performance of completion handler
Avoid the ref count gear in mlx5_cq_completion and mlx5_cq_event since it as
not required. The handlers need to be protected with rcu read locks while
inserts and deletes can safely occur concurrently with lookups (see
include/linux/radix-tree.h, Radix-tree synchronization).  When the CQ object is
deleted, we protect concurrent access to the radix tree with spinlock. After
that a call to synchronize_rcu will ensure there are no inflight handler
accessing the ojbect so we can safely free it.

Issue: 371296

Change-Id: I6f3f89b06f21d418c27622dffdcbea2a709d2db2
Signed-off-by: Eli Cohen <eli@mellanox.com>