net/mlx4: Add mlx4_bitmap zone allocator [topic: a0_v1_int]
The zone allocator is a mechanism which manages a few mlx4_bitmaps.

When allocating a resource, the user indicates the desired zone of
which this resource will be allocated from. If possible, the resource
will be allocated from this zone. Otherwise, the resource will be
allocated from a less-than, equal-to, higher-than priority zone,
according to the desired zone's properties with that respective
allocation order.

issue: 387689
Change-Id: I4fcdf003a12a299b36b67c35ba705b08271480f3
Signed-off-by: Matan Barak <matanb@mellanox.com>