net/mlx4_en: Support for configurable RSS hash function [topic: ethtool_rss_hash_function]
The ConnectX HW is capable of using one of the following hash functions:
Toeplitz and an XOR hash function.
This patch implements ethtool_ops {get,set}_rxfh_func callbacks for the
mlx4_en driver in order to query and configure the RSS hash function to
be used by the device.

Issue: 388542
Change-Id: I8ea5c5adb7615a4d92403ee8299f65e417be89b3
Signed-off-by: Eyal Perry <eyalpe@mellanox.com>