pgtable: Add API to query if write combining is available [topic: Add-is_writecombine_available]
Currently the kernel write-combining interface provides a best effort
mechanism in which the caller simply invokes pgprot_writecombine().

If write combining is available, the region is mapped for it, otherwise
the region is (silently) mapped as non-cached.

In some cases, however, the calling driver must know if write combining
is available, so a silent best effort mechanism is not sufficient.

Add writecombine_available(), which returns true if the system
supports write combining and false if it doesn't.

Issue: 427910
Change-Id: I32fbf10e652004724320a4f7ef50ee038f31ea0f
Signed-off-by: Moshe Lazer <moshel@mellanox.com>
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>

Conflicts:
	arch/arm64/include/asm/pgtable.h
	include/asm-generic/pgtable.h