net/mlx4_core: Replace mlx4 singlethread workqueue with system_wq [topic: nohz_full_workqueues]
A singlethread workqueue is unbounded by definition, therefore it could
harm a low-latency user application that is being run on a tickless CPU
(specified on a CPU list assigned to nohz_full argument in the kernel
cmdline, and nr_running <= 1). When it sets a work to be processed over
such CPU would cause nr_running > 1, and trigger the CPU ticks again.

Issue: 361059
Change-Id: Id722dffc1827c63cbb8c819da465b36106d82fc9
Signed-off-by: Eyal Perry <eyalpe@mellanox.com>