mlx4: Add A0 hybrid steering [topic: a0_v1_int]
A0 hybrid steering is a form of limited high performance flow
steering. By using this mode, mlx4 supported cards uses a
fast limited table based steering, in order to enable
fast steering of unicast packets to a QP.

For steering purposes, the usage of this hardware imposes
several limitations:
(1) Currently, flow steering rules with VLANs can't be used.
(2) A MAC is associated to exactly one QP.
(3) A QP is associated with exactly one MAC.
(4) The possible steerable QP range is very limited.

Because of limit (4), we use a zone based allocator.
We device our QPs to the following zones:
(1) General range
(2) Special MAC-assigned QPs [RSS, Raw-Ethernet] Each has its own
region.
    (a) RSS
    (b) Raw Ethernet

When we create a RSS QP or a raw Ethernet QP, we try hard to
allocate the QP from range (2). Otherwise, we try hard not to
allocate from this range. However, when the system is pushed
to its limits and one needs every resource, the allocator uses
every region it can.

Meaning, when we run out of raw-eth qps, the allocator allocates
from the general range (and the special-A0 area is no longer
active). If we run out of RSS qp, the mechanism tries to allocate from the
raw-eth QP zone. If that is also exhausted, the allocator will allocate from
the general range (and the A0 region is no longer active).

Note that if a raw Ethernet qp is allocated from the general range,
it attempts to allocate the range such that bits 6 and 7 (blueflame
bits) in the QP number are not set.

There are a few A0 suppoerted modes:
(a) Default mode - let the FW decide the flow steering mode
(b) Dynamic mode - FW will start in A0 and fallback to DMFS when
                   it can't fullfil the requried rules.
(c) Static mode - Only A0 complied rules are supported.

Apart from configuring and querying the FW, the different modes are
seamless to the drivers.

Changes from V0:
===============
- Squash the first 3 patches
- Remove 'net/mlx4: Add a check if there are too many reserved QPs'
- Squah some fixes from MLNX_OFED:
	- Don't crash if cleanup is called twice
- Fixed spelling mistakes / rephrasing
- Remove IPoIB from documentation

Change-Id: Ie5db4a4bb0f6398c2acf44ea4b7872623f5bd1fc