IB/core: Umem tunneling peer memory APIs [topic: peer_direct]
Builds umem over peer memory client functionality.
It tries getting a peer client for a given address range, in case found
further memory calls are tunneled to that peer client.
ib_umem_get was extended to have an indication whether this umem can
be part of a peer client. As a result, usage of
ib_umem_get was updated accordingly.

issue: 418558
Change-Id: I00802f115acd87d475e1875606074b92b5bfb2dc
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>