IB/core: Sysfs support for peer memory [topic: peer_direct]
Supplies the required functionality to expose information and
statistics over sysfs for a given peer memory client.

This mechanism enables userspace application to check
which peers are available (based on name & version) and based on that
decides whether it can run successfully.

Root sysfs directory is /sys/kernel/infiniband/<peer_name>, under that directory
will reside some files that represent the statistics for that peer.

issue: 418558
Change-Id: Ia760f9c6f867ac764e9f0e4013b9e23e75ef83b4
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>