net/mlx4_core: Relieve cpu load average on the port sending flow [net-next:bug-fix]
When a port is not attached, the FW requires a longer than usual time to
execute the SENSE_PORT command. In the command flow, the
wait_for_completion_timeout call used in mlx4_cmd_wait puts the kernel
thread into the uninterruptible state during this time. This, in turn,
due to the computation method, causes the CPU load average to increase.

Fix this by using wait_for_completion_interruptible_timeout() for the
SENSE_PORT command, which puts the thread in the interruptible state.
In this state, the thread does not contribute to the CPU load average.

Treat the interrupted case as if the SENSE_PORT command returned
port_type = NONE.

Fix suggested by Gideon Naim of Mellanox, and by Bart Van Assche.

issue: 489937

Change-Id: I5798d9686d878f82453a61567be8136fb47d208f
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>