net/mlx4_core: Restore port types as they were following internal error reset [topic: reset_flow]
HCAs are reset following internal errors, this commit restores the HCA to the
existing port link-layer settings (ib, eth, or auto-sense).  To enable this
restoration we manage a persistent area that is saved cross load/unload. If
more persistent data is required, new fields may be added to the persistent
data area.

issue: none
Change-Id: I8b2844ee8998423bfbdf69d29d70aba963cc5def
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>