net/mlx4_core: Set device configuration data to be persistent across reset [topic: reset_flow]
When an HCA enters an internal error state, this is detected by the driver.
The driver then should reset the HCA and restart the software stack. This
patch keeps ports information and some SRIOV configuration in a persistent area
to have it valid across reset.

If more persistent data is required, new fields may be added to the persistent
data area.

issue: 418559
Change-Id: I8b2844ee8998423bfbdf69d29d70aba963cc5def
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>