net/mlx4_core: Restore port types as they were following internal error [topic: reset_flow]
When an HCA enters an internal error state, this is detected by the driver.
The driver then should reset the HCA and restart the software stack. This
patch keeps current port types and after restarting the software stack restores
the HCA to the port types (ib, eth, or auto-sense) which existed before the
error occurred.

To enable this restoration we manage a persistent area that is saved across
load/unload.  If more persistent data is required, new fields may be added to
the persistent data area.

issue: 418559
Change-Id: I8b2844ee8998423bfbdf69d29d70aba963cc5def
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>