net/mlx4_core: manage interface state for Reset flow cases [topic: reset_flow]
We need to manage interface state to sync between reset flow and some other
relative cases such as remove_one. This has to be done to prevent certain
races. For example in case software stack is down as a result of unload call
the remove_one should skip the unload phase.

This patch implements the remove_one case, subsequent patches will handle the
AER and other cases.

The interface can be up/down, upon remove_one the state will include an extra
bit indicates that the device is cleaned-up forcing other tasks to be finish
before the final cleanup.

issue: 418559
Change-Id: I1caaa1a9a5aa96964143a27e2a02022ed56926c4
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>