net/mlx4_core: Reset flow activation upon SRIOV fatal command cases [topic: reset_flow]
When SRIOV commands are executed over the comm-channel and get
a fatal error (e.g. timeout, closing command failure) the VF enters
into error state and reset flow is activated.

To be able to recognize whether the failure was on a closing command
the operational code for the given VHCR command is used.
Once the device entered into an error state we prevent redundant
error messages from being printed.

issue: 418559
Change-Id: Icec816d629f480e30c7e73c69ca4dab010e5cc61
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>