IB/ipoib: Update broadcast record values after each successful join request [ib-next:ipoib-s1]
Update the cached broadcast record in the priv object after every new
join of this broadcast domain group.

These values are important for the port configuration (MTU size) and to
all the new join requests initial parameters.

For example, SM starts with 2K MTU for all the fabric, and after that it
restarts (or handover to new SM) with new port configuration of 4K MTU.
Without using the new values, the driver will keep its old configuration
of 2K and will not apply the new configuration of 4K.

issue: 420429
Change-Id: I51f0cdac8ef0091ce7767f97e882376acfde7eea
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>