IB/ipoib:Update broadcast record values after each success of join request to [ib-next:ipoib-s1] that group
Update the cached broadcast record in the priv object after every new
join of this broadcast domain group.

These values are important for the port configuration (MTU size) and to
all the new join requests initial parameters.

For example, sm starts with 2K MTU for all the fabric, and after that it
restarts (or handover to new sm) with new port configuration of 4K MTU,
without taking the new values the port will keep its old value (2K) and
will not move to 4K.

issue: 420429

Change-Id: Ic6b1329b178568073ff2eb3355e173c04bbd833a
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>