net/mlx4_core: Activate reset flow upon fatal command cases [topic: reset_flow]
We activate reset flow upon command fatal errors, when happened the device is
in a bad state and must be reset.

Below cases assumed to be fatal: FW command timed-out, an error from FW on
closing commands, pci is offline once posting/pending a command.  In those
cases we enter the device into an error state: chip is reset, pending commands
are waked-up and finished immediately, future commands will be immediately
returned.

The return code in above cases depends on the command, in case it's a closing
command which free and close resources the return value will be a success as
the chip was really reset, letting callers free their kerenl resources and
prevent memory leaks, other commands will return with -EIO.

As the device marked to be on an error state the catas poller will recognize
this case and restart software stuck for that device as done when FW internal
error was reported on its error buffer.  This flow must be done from the catas
flow and not from the command flow otherwise it might dead-lock pending that
command to be terminated.

issue: none
Change-Id: Id70a39c67e9d0b735e2a23807c12073fc631a815
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>