net/mlx4_core: Enable device recovery flow with SRIOV [topic: reset_flow]
In SRIOV, both the PF and the VF may attempt device recovery whenever it
assumes that the device is not functioning.  When the PF driver resets the
device, the VF driver should detect this and attempt to reinitialize the VF.
The VF driver must be able to reset the VF under all circumstances, even if the
PF driver is not responsive.

The VF driver shall reset its VF in the following cases:

1. Commands are not processed within reasonable time over the communication channel.
This case is handled in next patch of that series, considering device state
and the correct return code based on the command as was done in the native mode.

2. The VF driver receives an internal error event reported by the PF on the
communication channel. This occurs when the PF driver resets the device or when
VF is out of sync with the PF.

This patch adds 'VF reset' capability, which allows the VF driver to
reinitialize the VF and its driver even when PF driver is not responsive.

As PF and VF may run their reset flow simulantanisly there are several cases
that are handled:
- Prevent freeing VF resources upon FLR when PF is in its unloading stage.
- Prevent PF getting VF commands before it has finished initializing its resources.
- Upon VF startup, it checks that comm-channel is online before sending
  commands to the PF and getting timed-out.

issue: 418559
Change-Id: I3f216150765d07a622e62bbb59b30f809e1e1ba2
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>