IB/core: notify rdma_cm clients when device is removed
With InfiniBand, RDMA_CM applications are being notified when the
ib_device is being removed and can react by closing resources.
In the same way, with RoCE, presence of a net_device is mandatory
for Verbs applications to work. Therefore, We want that RDMA_CM
application will be notified when the net device is being removed.

issue: 515968
Change-Id: I732bdfa0173e0c036b4ef98edf7978239891ee7d
Signed-off-by: Moni Shoua <monis@mellanox.com>